Improve: Update styles and add sections
