+++ src/app/sections/backoffice-features.tsx
@@ -1,0 +1,150 @@
+"use client";
+
+import { Lock, ClockArrowUp, Sparkles, Cog, ShoppingCart } from "lucide-react";
+import RotateCard from "@/components/animations/rotate-card";
+import { useState } from "react";
+import Image from "next/image";
+import { Container } from "@/components/container/container";
+import Shadow from "@/components/shadow/shadow";
+
+type Feature = {
+  id: string;
+  title: string;
+  description: string;
+  icon: React.ElementType;
+};
+
+const BACKOFFICE_FEATURES: Feature[] = [
+  {
+    id: "gestionar-carta",
+    title: "Gestionar la carta",
+    description:
+      "Modifique precios, unidades, alérgenos, disponibilidad, sugerencias del chef y muchas más opciones",
+    icon: Lock,
+  },
+  {
+    id: "gestionar-videos",
+    title: "Gestionar los videos",
+    description:
+      "Modifique los videos de los platos para que sean más atractivos",
+    icon: Lock,
+  },
+  {
+    id: "gestionar-comentarios",
+    title: "Gestionar los comentarios",
+    description: "Lea los comentarios de los clientes y responda a sus dudas",
+    icon: ClockArrowUp,
+  },
+  {
+    id: "gestionar-sugerencias",
+    title: "Gestionar las sugerencias",
+    description: "Lea las sugerencias de los clientes y responda a sus dudas",
+    icon: Sparkles,
+  },
+  {
+    id: "gestionar-estadisticas",
+    title: "Gestionar las estadísticas",
+    description: "Acceda a las estadísticas de la carta y de los clientes",
+    icon: Cog,
+  },
+  {
+    id: "gestionar-notificaciones",
+    title: "Gestionar las notificaciones",
+    description: "Reciba notificaciones de los clientes y responda a sus dudas",
+    icon: ShoppingCart,
+  },
+];
+
+interface BackofficeFeatureCardProps {
+  title: string;
+  description: string;
+  icon: React.ElementType;
+}
+
+const BackofficeFeaturesSection: React.FC = () => {
+  const [showVideo, setShowVideo] = useState(false);
+
+  return (
+    <Container id="caracteristicas" className="flex flex-col">
+      <div className="mx-auto max-w-7xl px-6 lg:px-8">
+        <div className="mx-auto max-w-3xl sm:text-center">
+          <h2 className="text-base/7 font-semibold text-primary-400">
+            Todo lo que necesitas para gestionar tu carta
+          </h2>
+          <p className="mt-2 text-4xl font-semibold tracking-tight text-pretty sm:text-5xl sm:text-balance">
+            Un panel de gestión para actualizar la carta en tiempo real
+          </p>
+          <p className="my-8 text-lg/8 muted">
+            Modifique precios, unidades, alérgenos, disponibilidad, sugerencias
+            del chef y muchas más opciones desde un solo lugar y en tiempo real
+          </p>
+        </div>
+      </div>
+      <RotateCard height={800}>
+        {showVideo ? (
+          <video
+            autoPlay
+            loop
+            muted
+            controls={false}
+            src="/demo.webm"
+            width={1200}
+            height={800}
+            className="rounded-xl ring-1 shadow-2xl ring-white/10 object-cover border-10 border-black/50"
+          />
+        ) : (
+          <div
+            className="relative group cursor-pointer"
+            onClick={() => setShowVideo(true)}
+          >
+            <Image
+              src="/images/backoffice-dark.webp"
+              alt="Demo"
+              width={1200}
+              height={800}
+              className="rounded-xl ring-1 shadow-2xl ring-white/10 object-cover border-10 border-black/50"
+            />
+            <div className="absolute inset-0 flex items-center justify-center bg-black/60 opacity-0 group-hover:opacity-100 transition-opacity duration-300 rounded-xl">
+              <h1 className="text-white text-2xl font-bold">Visualizar demo</h1>
+            </div>
+          </div>
+        )}
+      </RotateCard>
+
+      <div className="mx-auto mt-16 max-w-7xl px-6 sm:mt-20 md:mt-24 lg:px-8">
+        <dl className="mx-auto grid max-w-2xl grid-cols-1 gap-x-6 gap-y-10 text-base/7 sm:grid-cols-2 lg:mx-0 lg:max-w-none lg:grid-cols-3 lg:gap-x-8 lg:gap-y-8">
+          <Shadow />
+          {BACKOFFICE_FEATURES.map((feature) => (
+            <>
+              <BackofficeFeatureCard
+                key={feature.id}
+                title={feature.title}
+                description={feature.description}
+                icon={feature.icon}
+              />
+            </>
+          ))}
+        </dl>
+      </div>
+    </Container>
+  );
+};
+
+function BackofficeFeatureCard({
+  title,
+  description,
+  icon: Icon,
+}: BackofficeFeatureCardProps) {
+  return (
+    <div className="flex flex-col items-start gap-4 rounded-lg p-6 border border-primary/90 glass-effect">
+      <div className="mb-2 flex h-12 w-12 items-center justify-center rounded-full bg-primary-400/20 text-primary-400 shadow-lg shadow-primary/10 backdrop-blur-md">
+        <Icon className="h-6 w-6 text-primary-dark" />
+      </div>
+      <dt className="text-lg font-semibold text-white dark:text-gray-100 drop-shadow">
+        {title}
+      </dt>
+      <dd className="text-base text-white dark:text-gray-300">{description}</dd>
+    </div>
+  );
+}
+export default BackofficeFeaturesSection;
