+++ src/app/sections/stats.tsx
@@ -1,0 +1,29 @@
+import { Container } from "@/components/container/container";
+
+const stats = [
+  { id: 1, name: "Incremento medio de ingresos", value: "+18%" },
+  { id: 2, name: "Restaurantes que confían en MenuVision", value: "250+" },
+  { id: 3, name: "Cartas digitales creadas", value: "1.200+" },
+];
+
+export default function StatsSection() {
+  return (
+    <Container className="py-24 sm:py-32">
+      <div className="mx-auto max-w-7xl px-6 lg:px-8">
+        <dl className="grid grid-cols-1 gap-x-8 gap-y-16 text-center lg:grid-cols-3">
+          {stats.map((stat) => (
+            <div
+              key={stat.id}
+              className="mx-auto flex max-w-xs flex-col gap-y-4"
+            >
+              <dt className="text-base/7 text-gray-600">{stat.name}</dt>
+              <dd className="order-first text-3xl font-semibold tracking-tight text-gray-900 sm:text-5xl">
+                {stat.value}
+              </dd>
+            </div>
+          ))}
+        </dl>
+      </div>
+    </Container>
+  );
+}
