--- src/sections/call-to-action.tsx
@@ -1,75 +1,0 @@
-import { CheckCircleIcon } from "@heroicons/react/20/solid";
-import Link from "next/link";
-
-const benefits = [
-  "Diseño personalizado",
-  "Panel de control",
-  "Gestión de platos",
-  "Gestión de categorías",
-  "Gestión de imágenes",
-  "Gestión de alérgenos",
-  "Gestión de preferencias",
-  "Gestión de platos recomendados",
-  "Gestión de platos visibles o no",
-];
-
-export default function CallToActionSection() {
-  return (
-    <div className="overflow-hidden bg-gray-900 py-24 sm:py-32">
-      <div className="relative isolate">
-        <div className="mx-auto max-w-7xl sm:px-6 lg:px-8">
-          <div className="mx-auto flex max-w-2xl flex-col gap-16 bg-white/3 px-6 py-16 ring-1 ring-white/10 sm:rounded-3xl sm:p-8 lg:mx-0 lg:max-w-none lg:flex-row lg:items-center lg:py-20 xl:gap-x-20 xl:px-20">
-            <img
-              alt=""
-              src="/images/team.webp"
-              className="h-96 w-full flex-none rounded-2xl object-cover shadow-xl lg:aspect-square lg:h-auto lg:max-w-sm"
-            />
-            <div className="w-full flex-auto">
-              <h2 className="text-4xl font-semibold tracking-tight text-pretty text-white sm:text-5xl">
-                Contacta con nosotros
-              </h2>
-              <p className="mt-6 text-lg/8 text-pretty text-gray-400">
-                Si tienes alguna duda, no dudes en contactar con nosotros.
-              </p>
-              <ul
-                role="list"
-                className="mt-10 grid grid-cols-1 gap-x-8 gap-y-3 text-base/7 text-gray-200 sm:grid-cols-2"
-              >
-                {benefits.map((benefit) => (
-                  <li key={benefit} className="flex gap-x-3">
-                    <CheckCircleIcon
-                      aria-hidden="true"
-                      className="h-7 w-5 flex-none text-gray-200"
-                    />
-                    {benefit}
-                  </li>
-                ))}
-              </ul>
-              <div className="mt-10 flex">
-                <Link
-                  href="#contact-us"
-                  className="text-sm/6 font-semibold text-primary-dark hover:text-primary-300"
-                >
-                  Contacta con nosotros
-                  <span aria-hidden="true">&rarr;</span>
-                </Link>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div
-          aria-hidden="true"
-          className="absolute inset-x-0 -top-16 -z-10 flex transform-gpu justify-center overflow-hidden blur-3xl"
-        >
-          <div
-            style={{
-              clipPath:
-                "polygon(73.6% 51.7%, 91.7% 11.8%, 100% 46.4%, 97.4% 82.2%, 92.5% 84.9%, 75.7% 64%, 55.3% 47.5%, 46.5% 49.4%, 45% 62.9%, 50.3% 87.2%, 21.3% 64.1%, 0.1% 100%, 5.4% 51.1%, 21.4% 63.9%, 58.9% 0.2%, 73.6% 51.7%)",
-            }}
-            className="aspect-1318/752 w-329.5 flex-none bg-linear-to-r from-[#80caff] to-[#4f46e5] opacity-20"
-          />
-        </div>
-      </div>
-    </div>
-  );
-}
