--- src/sections/contact-us.tsx
@@ -1,220 +1,0 @@
-"use client";
-import {
-  BuildingOffice2Icon,
-  EnvelopeIcon,
-  PhoneIcon,
-} from "@heroicons/react/24/outline";
-import { useState } from "react";
-import { usePostTicket } from "@/api/use-post-ticket";
-
-export default function ContactUsSection() {
-  const { postTicket, isLoading } = usePostTicket();
-  const [formData, setFormData] = useState({
-    source: "LANDING",
-    name: "",
-    surname: "",
-    email: "",
-    message: "",
-    subject: "",
-  });
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const response = await postTicket(formData);
-    if (response) {
-      console.log(response);
-    }
-  };
-
-  return (
-    <div className="relative isolate" id="contacto">
-      <div className="mx-auto grid max-w-7xl grid-cols-1 lg:grid-cols-2">
-        <div className="relative px-6 pt-24 pb-20 sm:pt-32 lg:static lg:px-8 lg:py-48">
-          <div className="mx-auto max-w-xl lg:mx-0 lg:max-w-lg">
-            <div
-              aria-hidden="true"
-              className="absolute inset-x-0 -top-16 -z-10 flex transform-gpu justify-center overflow-hidden blur-3xl"
-            >
-              <div
-                style={{
-                  clipPath:
-                    "polygon(73.6% 51.7%, 91.7% 11.8%, 100% 46.4%, 97.4% 82.2%, 92.5% 84.9%, 75.7% 64%, 55.3% 47.5%, 46.5% 49.4%, 45% 62.9%, 50.3% 87.2%, 21.3% 64.1%, 0.1% 100%, 5.4% 51.1%, 21.4% 63.9%, 58.9% 0.2%, 73.6% 51.7%)",
-                }}
-                className="aspect-1318/752 w-329.5 flex-none bg-linear-to-r from-[#14b8a6] to-[#2dd4bf] opacity-20"
-              />
-            </div>
-            <h2 className="text-4xl font-semibold tracking-tight text-pretty text-gray-900 sm:text-5xl">
-              Contacta con nosotros
-            </h2>
-            <p className="mt-6 text-lg/8 text-gray-600">
-              ¿Quieres llevar tu restaurante al siguiente nivel con una carta
-              digital QR? Estamos aquí para ayudarte a implementar la mejor
-              solución para tu negocio.
-            </p>
-            <dl className="mt-10 space-y-4 text-base/7 text-gray-600">
-              <div className="flex gap-x-4">
-                <dt className="flex-none">
-                  <span className="sr-only">Dirección</span>
-                  <BuildingOffice2Icon
-                    aria-hidden="true"
-                    className="h-7 w-6 text-gray-400"
-                  />
-                </dt>
-                <dd>
-                  Madrid, España
-                  <br />
-                  Disponible en toda España
-                </dd>
-              </div>
-              <div className="flex gap-x-4">
-                <dt className="flex-none">
-                  <span className="sr-only">Teléfono</span>
-                  <PhoneIcon
-                    aria-hidden="true"
-                    className="h-7 w-6 text-gray-400"
-                  />
-                </dt>
-                <dd>
-                  <a href="tel:+34 600 000 000" className="hover:text-gray-900">
-                    +34 600 000 000
-                  </a>
-                </dd>
-              </div>
-              <div className="flex gap-x-4">
-                <dt className="flex-none">
-                  <span className="sr-only">Email</span>
-                  <EnvelopeIcon
-                    aria-hidden="true"
-                    className="h-7 w-6 text-gray-400"
-                  />
-                </dt>
-                <dd>
-                  <a
-                    href="mailto:info@menuvision.es"
-                    className="hover:text-gray-900"
-                  >
-                    info@menuvision.es
-                  </a>
-                </dd>
-              </div>
-            </dl>
-          </div>
-        </div>
-        <form
-          onSubmit={handleSubmit}
-          className="px-6 pt-20 pb-24 sm:pb-32 lg:px-8 lg:py-48"
-        >
-          <div className="mx-auto max-w-xl lg:mr-0 lg:max-w-lg">
-            <div className="grid grid-cols-1 gap-x-8 gap-y-6 sm:grid-cols-2">
-              <div>
-                <label
-                  htmlFor="name"
-                  className="block text-sm/6 font-semibold text-gray-900"
-                >
-                  Nombre
-                </label>
-                <div className="mt-2.5">
-                  <input
-                    id="name"
-                    name="name"
-                    type="text"
-                    autoComplete="given-name"
-                    className="block w-full rounded-md bg-white px-3.5 py-2 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-primary"
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-              <div>
-                <label
-                  htmlFor="surname"
-                  className="block text-sm/6 font-semibold text-gray-900"
-                >
-                  Apellidos
-                </label>
-                <div className="mt-2.5">
-                  <input
-                    id="surname"
-                    name="surname"
-                    type="text"
-                    autoComplete="family-name"
-                    className="block w-full rounded-md bg-white px-3.5 py-2 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-primary"
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-              <div className="sm:col-span-2">
-                <label
-                  htmlFor="email"
-                  className="block text-sm/6 font-semibold text-gray-900"
-                >
-                  Correo electrónico
-                </label>
-                <div className="mt-2.5">
-                  <input
-                    id="email"
-                    name="email"
-                    type="email"
-                    autoComplete="email"
-                    className="block w-full rounded-md bg-white px-3.5 py-2 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-primary"
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-
-              <div className="sm:col-span-2">
-                <label
-                  htmlFor="subject"
-                  className="block text-sm/6 font-semibold text-gray-900"
-                >
-                  Asunto
-                </label>
-                <div className="mt-2.5">
-                  <input
-                    id="subject"
-                    name="subject"
-                    className="block w-full rounded-md bg-white px-3.5 py-2 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-primary"
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-
-              <div className="sm:col-span-2">
-                <label
-                  htmlFor="message"
-                  className="block text-sm/6 font-semibold text-gray-900"
-                >
-                  Mensaje
-                </label>
-                <div className="mt-2.5">
-                  <textarea
-                    id="message"
-                    name="message"
-                    rows={4}
-                    className="block w-full rounded-md bg-white px-3.5 py-2 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-primary"
-                    defaultValue={""}
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-            </div>
-            <div className="mt-8 flex justify-end">
-              <button
-                type="submit"
-                disabled={isLoading}
-                className="rounded-md bg-primary-dark px-3.5 py-2.5 text-center text-sm font-semibold text-white shadow-xs hover:bg-primary focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
-              >
-                {isLoading ? "Enviando..." : "Enviar mensaje"}
-              </button>
-            </div>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
